+++ protected/humhub/modules/content/Events.php
@@ -79,14 +79,15 @@ class Events extends BaseObject
 
         $integrityController->showTestHeadline('Content Objects (' . Content::find()->count() . ' entries)');
         foreach (Content::find()->each() as $content) {
+            /* @var Content $content */
             if ($content->createdBy == null) {
                 if ($integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid user!')) {
-                    $content->delete();
+                    $content->hardDelete();
                 }
             }
             if ($content->getPolymorphicRelation() === null) {
                 if ($integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid content object!')) {
-                    $content->delete();
+                    $content->hardDelete();
                 }
             }
         }